Correct Repository Failure
When cloning for the first time: It can't pull down the "gcc4mbed".
GitHub say’s Permissions failed. Since it fails, there is a change in
the "Uncommitted Changes" of the "gcc4mbed" being deleted and well….it
won’t sync with uncommitted changes, and won’t discard the changes so
the only resolution is to commit this change.